requirements.txt
torchvision
matplotlib
torch
pillow
tqdm
